vl-foreldrepenger/domenetjenester/beregning-ytelse/regelmodell/src/main/java/no/nav/foreldrepenger/beregning/regelmodell/beregningsgrunnlag/Beregningsgrunnlag.java
package no.nav.foreldrepenger.beregning.regelmodell.beregningsgrunnlag;

import java.time.LocalDate;
import java.util.ArrayList;
import java.util.List;
import java.util.Objects;

import com.fasterxml.jackson.annotation.JsonManagedReference;

public class Beregningsgrunnlag {

    private LocalDate skjæringstidspunkt;
    private final List<AktivitetStatus> aktivitetStatuser = new ArrayList<>();
    @JsonManagedReference
    private final List<BeregningsgrunnlagPeriode> beregningsgrunnlagPerioder = new ArrayList<>();

    private Beregningsgrunnlag() {
    }

    public List<BeregningsgrunnlagPeriode> getBeregningsgrunnlagPerioder() {
        return beregningsgrunnlagPerioder;
    }

    public static Builder builder() {
        return new Builder();
    }

    public static class Builder {
        private Beregningsgrunnlag beregningsgrunnlagMal;

        public Builder() {
            beregningsgrunnlagMal = new Beregningsgrunnlag();
        }

        public Builder medSkjæringstidspunkt(LocalDate skjæringstidspunkt){
            beregningsgrunnlagMal.skjæringstidspunkt = skjæringstidspunkt;
            return this;
        }

        public Builder medAktivitetStatuser(List<AktivitetStatus> aktivitetStatusList){
            beregningsgrunnlagMal.aktivitetStatuser.addAll(aktivitetStatusList);
            return this;
        }

        public Builder medBeregningsgrunnlagPeriode(BeregningsgrunnlagPeriode beregningsgrunnlagPeriode) {
            beregningsgrunnlagMal.beregningsgrunnlagPerioder.add(beregningsgrunnlagPeriode);
            return this;
        }

        public Builder medBeregningsgrunnlagPerioder(List<BeregningsgrunnlagPeriode> beregningsgrunnlagPerioder) {
            beregningsgrunnlagMal.beregningsgrunnlagPerioder.addAll(beregningsgrunnlagPerioder);
            return this;
        }

        public Beregningsgrunnlag build() {
            verifyStateForBuild();
            return beregningsgrunnlagMal;
        }

        void verifyStateForBuild() {
            Objects.requireNonNull(beregningsgrunnlagMal.skjæringstidspunkt, "skjæringstidspunkt");
            Objects.requireNonNull(beregningsgrunnlagMal.aktivitetStatuser, "aktivitetStatuser");
            if (beregningsgrunnlagMal.beregningsgrunnlagPerioder.isEmpty()) {
                throw new IllegalStateException("Beregningsgrunnlaget må inneholde minst 1 periode");
            }
            if (beregningsgrunnlagMal.aktivitetStatuser.isEmpty()) {
                throw new IllegalStateException("Beregningsgrunnlaget må inneholde minst 1 status");
            }
        }
    }
}
